+++ src/Client/ui/pages/arena.jsx
@@ -6,29 +6,28 @@ import HandColumn from "Client/ui/components/fragments/hand-column";
 import ControlColumn from "Client/ui/components/fragments/control-column";
 import StatusColumn from "Client/ui/components/fragments/status-column";
 import DeckActions from "Client/actions/deck";
-import DestroyActions from "Client/actions/destroy";
-import OutActions from "Client/actions/out"
-import HandActions from "Client/actions/hand";
 import MenuActions from "Client/actions/menu";
-import BoardActions from "Client/actions/board";
+import Moves from "Client/moves";
 import Card from "Client/ui/components/card/card";
-import Strings from "utils/strings";
 import "Client/ui/styles/arena.css";
 
 const Arena = ({G, ctx, playerID, deckID, moves, events}) => {
   const myID = parseInt(playerID);
   const rivalID = myID === 0 ? 1 : 0;
 
-  const {getBoardActionMenu, getTileCardsList, moveInBoard} = BoardActions
-  const {getDZForSearch, reborn} = DestroyActions
-  const {spawnFaceUp, spawnFaceDown, getHandActionsOnMenu} = HandActions
-  const {getOOGForSearch} = OutActions
-  const {constructDeck, getDeckActionsOnMenu, getDeckForSearch} = DeckActions
   const [actionMenu, setActionMenu] = useState(null);
   const [highlightCard, setHighlightCard] = useState(Card("", -1));
   const [listMenu, setListMenu] = useState(null);
   const [lifeMenu, setLifeMenu] = useState(null);
   const [selectToBoard, setSelectToBoard] = useState(null);
+
+  const setters = {setActionMenu, setHighlightCard, setListMenu, setLifeMenu, setSelectToBoard}
+  const getters = {actionMenu, highlightCard, listMenu, lifeMenu, selectToBoard}
+
+  const params = {G, myID, moves, ...setters, ...getters,}
+
+  const {deckMenu, dzMenu, oogMenu, handMenu,
+    boardMenu, clearMenuCallback} = MenuActions(params);
 
   const isSelected = (place) =>
     selectToBoard && selectToBoard.origin[place] !== undefined;
@@ -36,89 +35,19 @@ const Arena = ({G, ctx, playerID, deckID, moves, events}) => {
   React.useEffect(() => {
     const deckStart = async () => {
       let cards = await getDeckService(deckID);
-      moves.setDeck(constructDeck(deckID, cards, myID));
+      moves.setDeck(DeckActions.constructDeck(deckID, cards, myID));
     };
 
     if (deckID !== G.deck[myID].id) {
       deckStart();
     }
-  }, [constructDeck, moves, deckID, G, myID]);
+  }, [moves, deckID, G, myID]);
 
-  //TODO REMOVE THIS AFTER
-  const deckMenu = (e) => {
-    setSelectToBoard(null);
-    setActionMenu(getDeckActionsOnMenu(e));
-  };
-
-  const dzMenu = (id, mine) => {
-    setListMenu(getDZForSearch(G, id, mine));
-  };
-
-  const oogMenu = () => {
-    setListMenu(getOOGForSearch(G.out));
-  };
-
-  const handMenu = (e, i) => {
-    setSelectToBoard(null);
-    setActionMenu(getHandActionsOnMenu(e, i, G.hand, myID));
-  };
-
-  const boardMenu = (e, tile, id) => {
-    setActionMenu(getBoardActionMenu(e, tile, id));
-  };
-
-  const setMenu = (menu) => {
-    setActionMenu(menu);
-  };
-
-  const clearListMenu = () => {
-    let wasDeckMenu = listMenu.header === Strings.deckHeader;
-    setListMenu(null);
-    if (wasDeckMenu) {
-      setTimeout(moves.shuffleDeck, 100);
-    }
-  };
-
-  const clearMenuCallback = () => {
-    if (actionMenu) {
-      setActionMenu(null);
-    } else if (listMenu) {
-      clearListMenu();
-    } else if (lifeMenu) {
-      setLifeMenu(null);
-    } else if (G.reveal[myID]) {
-      moves.clearReveal();
-    }
-  };
- //TODO @END
   const clearSelectionCallback = () => {
     setSelectToBoard(null);
   };
 
-  const clientSideMoves = {
-    spawnFaceUp: (...args) => {
-      setSelectToBoard(spawnFaceUp(...args));
-    },
-    spawnFaceDown: (...args) => {
-      setSelectToBoard(spawnFaceDown(...args));
-    },
-    reborn: (...args) => {
-      setSelectToBoard(reborn(...args));
-    },
-    getDeckForSearch: () => {
-      setListMenu(getDeckForSearch(G.deck[myID].cards));
-    },
-    myLifeMenu: () => {
-      setLifeMenu(MenuActions.getLifeMenu(G.life[myID]));
-    },
-    setMenu,
-    getTileCardsList: (...args) => {
-      setListMenu(getTileCardsList(...args))
-    },
-    moveInBoard: (...args) => {
-      setSelectToBoard(moveInBoard(G.board, ...args));
-    },
-  };
+  const clientSideMoves = Moves(params);
 
   return (
     <div className="arena">